+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   * Fixed gravity measurement bug from IMU sensor
   * All sensors are now multi-stream, that means that the same sensor can be listened from different clients
   * Fixed point cloud of LiDAR. Now the points are given correctly in the sensor's coordinate system.
+  * Fixed light intensity and camera parameters to match
   * Exposed matrix form of transformation to the client and Python API.
   * Added make command to download contributions as plugins (`make plugins`)
   * Added PythonAPI command to set multiple car light states at once